+++ src/common/plugins/doughnutLabel.js
@@ -6,7 +6,10 @@ const TextColor =
 export default {
   id: 'doughnutLabel',
   beforeDraw: (chart, args, options) => {
-    if (chart.config.type === 'doughnut') {
+    if (
+      chart.config.type === 'doughnut' &&
+      !chart.config.options.doughnutLabel?.disabled
+    ) {
       const { ctx } = chart;
 
       // get sum of all visible data points
@@ -22,12 +25,32 @@ export default {
       ctx.fillStyle = TextColor;
 
       //Draw text in center
-      ctx.fillText(total, centerX, centerY);
-      ctx.fillText(
-        chart.config.options.scales.y.title.text,
-        centerX,
-        centerY + 20
-      );
+      // get custom options for center text
+      const Line1textOption = chart.config.options.doughnutLabel?.line1text;
+      const Line2textOption = chart.config.options.doughnutLabel?.line2text;
+
+      // set default values
+      let Line1text = total;
+      let Line2text = chart.config.options.scales.y.title.text;
+
+      // determine if line1text option is given, is a function or not, and update text
+      if (Line1textOption || Line1textOption === '') {
+        Line1text =
+          typeof Line1textOption === 'function'
+            ? Line1textOption(Line1text, ctx)
+            : Line1textOption;
+      }
+
+      // determine if line2text option is given, is a function or not, and update text
+      if (Line2textOption || Line2textOption === '') {
+        Line2text =
+          typeof Line2textOption === 'function'
+            ? Line2textOption(Line2text, ctx)
+            : Line2textOption;
+      }
+
+      ctx.fillText(Line1text, centerX, centerY);
+      ctx.fillText(Line2text, centerX, centerY + 20);
     }
   },
 };