+++ src/stories/Bubble.stories.js
@@ -79,6 +79,8 @@ const args = {
   },
   hideDescription: false,
   hideCaptions: false,
+  hideHeader: false,
+  hideControls: false,
   colorPalette: 'categorical',
   noBorder: false,
   width: null,
@@ -97,6 +99,8 @@ export const Bubble = {
         .datasets=${args.datasets}
         ?hideDescription=${args.hideDescription}
         ?hideCaptions=${args.hideCaptions}
+        ?hideHeader=${args.hideHeader}
+        ?hideControls=${args.hideControls}
         ?noBorder=${args.noBorder}
         .options=${{ colorPalette: args.colorPalette, ...args.options }}
         .width=${args.width}