+++ src/stories/Combo.stories.js
@@ -50,6 +50,8 @@ const args = {
   },
   hideDescription: false,
   hideCaptions: false,
+  hideHeader: false,
+  hideControls: false,
   colorPalette: 'categorical',
   noBorder: false,
   width: null,
@@ -68,6 +70,8 @@ export const Combo = {
         .datasets=${args.datasets}
         ?hideDescription=${args.hideDescription}
         ?hideCaptions=${args.hideCaptions}
+        ?hideHeader=${args.hideHeader}
+        ?hideControls=${args.hideControls}
         ?noBorder=${args.noBorder}
         .options=${{ colorPalette: args.colorPalette, ...args.options }}
         .width=${args.width}
@@ -125,6 +129,8 @@ export const MultiAxis = {
         .datasets=${args.datasets}
         ?hideDescription=${args.hideDescription}
         ?hideCaptions=${args.hideCaptions}
+        ?hideHeader=${args.hideHeader}
+        ?hideControls=${args.hideControls}
         ?noBorder=${args.noBorder}
         .options=${{ colorPalette: args.colorPalette, ...args.options }}
         .width=${args.width}
@@ -176,6 +182,8 @@ export const ComboStacked = {
         .datasets=${args.datasets}
         ?hideDescription=${args.hideDescription}
         ?hideCaptions=${args.hideCaptions}
+        ?hideHeader=${args.hideHeader}
+        ?hideControls=${args.hideControls}
         ?noBorder=${args.noBorder}
         .options=${{ colorPalette: args.colorPalette, ...args.options }}
         .width=${args.width}
@@ -233,6 +241,8 @@ export const ComboFloating = {
         .datasets=${args.datasets}
         ?hideDescription=${args.hideDescription}
         ?hideCaptions=${args.hideCaptions}
+        ?hideHeader=${args.hideHeader}
+        ?hideControls=${args.hideControls}
         ?noBorder=${args.noBorder}
         .options=${{ colorPalette: args.colorPalette, ...args.options }}
         .width=${args.width}