+++ src/stories/Geo.stories.js
@@ -36,6 +36,8 @@ const args = {
   description: 'Chart description.',
   hideDescription: false,
   hideCaptions: false,
+  hideHeader: false,
+  hideControls: false,
   colorPalette: 'sequential01',
   noBorder: false,
   width: null,
@@ -49,7 +51,10 @@ export const WorldChoropleth = {
     datasets: [
       {
         label: 'Countries',
-        data: countries.map((d) => ({ feature: d, value: Math.random() * 10 })),
+        data: countries.map((d) => ({
+          feature: d,
+          value: d.properties.name.length * 10,
+        })),
       },
     ],
   },
@@ -63,6 +68,8 @@ export const WorldChoropleth = {
         .datasets=${args.datasets}
         ?hideDescription=${args.hideDescription}
         ?hideCaptions=${args.hideCaptions}
+        ?hideHeader=${args.hideHeader}
+        ?hideControls=${args.hideControls}
         ?noBorder=${args.noBorder}
         .options=${{ colorPalette: args.colorPalette, ...args.options }}
         .width=${args.width}
@@ -80,7 +87,10 @@ export const CountryChoropleth = {
       {
         label: 'States',
         outline: nation,
-        data: states.map((d) => ({ feature: d, value: Math.random() * 10 })),
+        data: states.map((d) => ({
+          feature: d,
+          value: d.properties.name.length * 10,
+        })),
       },
     ],
     options: {
@@ -102,6 +112,8 @@ export const CountryChoropleth = {
         .datasets=${args.datasets}
         ?hideDescription=${args.hideDescription}
         ?hideCaptions=${args.hideCaptions}
+        ?hideHeader=${args.hideHeader}
+        ?hideControls=${args.hideControls}
         ?noBorder=${args.noBorder}
         .options=${{ colorPalette: args.colorPalette, ...args.options }}
         .width=${args.width}
@@ -120,7 +132,7 @@ export const WorldBubbleMap = {
         outline: countries,
         label: 'Countries',
         data: nationCapitals.map((d) =>
-          Object.assign(d, { value: Math.round(Math.random() * 100) })
+          Object.assign(d, { value: d.description.length * 100 })
         ),
       },
     ],
@@ -136,6 +148,8 @@ export const WorldBubbleMap = {
         .datasets=${args.datasets}
         ?hideDescription=${args.hideDescription}
         ?hideCaptions=${args.hideCaptions}
+        ?hideHeader=${args.hideHeader}
+        ?hideControls=${args.hideControls}
         ?noBorder=${args.noBorder}
         .options=${{ colorPalette: args.colorPalette, ...args.options }}
         .width=${args.width}
@@ -153,7 +167,7 @@ export const CountryBubbleMap = {
       {
         outline: states,
         data: capitals.map((d) =>
-          Object.assign(d, { value: Math.round(Math.random() * 100) })
+          Object.assign(d, { value: d.description.length * 100 })
         ),
       },
     ],
@@ -177,6 +191,8 @@ export const CountryBubbleMap = {
         .datasets=${args.datasets}
         ?hideDescription=${args.hideDescription}
         ?hideCaptions=${args.hideCaptions}
+        ?hideHeader=${args.hideHeader}
+        ?hideControls=${args.hideControls}
         ?noBorder=${args.noBorder}
         .options=${{ colorPalette: args.colorPalette, ...args.options }}
         .width=${args.width}